添加升级文件保存相关函数

+++ User/inc/utility.h
@@ -2,7 +2,7 @@
 #define __UTILITY_H
 
 void Uart_SendByte(LPUART_Type *base, uint8_t data);
-uint16_t CRC16(uint8_t *data,uint8_t length);
+uint16_t CRC16(uint8_t *data,uint32_t length);
 
 
 #endif
